ChatGPT/flask_apps/app1/backend/requirements.txt
Flask
flask-cors
pyjwt
bcrypt